tests/requirements.txt
pytest==7.4.2
pytest-playwright

ipdb
ipython
